Add files via upload
The Project titled "Online Crime Server" is a web-based application. This software provides all crime management solutions, which are easily accessible to everyone & facility for Reporting Online Crimes, Complaints, Missing Persons, and Most wanted, etc. No. of clients can connect to the server. The Online Crime Server Application starts with the common people who want to file a complaint through the website, so it can be very useful for the police department to find out the problem in society without people visiting the police station. Crime Management is a critical process having many calculations & operations. We ensure data accuracy, proper control of higher officials, greater efficiency, better service, user-friendly & interactive application, minimizing the manual data entry & time needed for various processing. Online Crime Management System is the Computerization of the present manual system of Police Stations.  The present system is manual. In the case of the present manual system, all the record is kept in a documented form and stored in various types of registers. The manual system has no proper system for record storage and storing information about crimes and criminals. The present system of police is not automated. The documentation of criminal records is done manually. Case files are dumped in heaps while some are kept on shelves. These important documents quickly get dusty and because of lack of proper storage, some criminal documents are seriously damaged. With this kind of situation, it becomes highly difficult and even impossible to track down known criminals, access the past criminal history of a suspect, and know the status of some concluded cases. In police stations, all the file work is done by the staff, so the burden of work leads to the criminal's information being delayed. There are also problems with record storage, updating, backup and recovery, and redundancy of data. To overcome the problems in the present system we design a computerized information system, which is free of all the flaws, present in the manual system of the institute